refactor(web): consolidate browse components and eliminate duplicate card wrappers
- Consolidate BrowseServices and BrowseSpareParts to render AdCardGrid directly
- Delete duplicate wrappers BrowseServicesCard.tsx and BrowseSparePartsCard.tsx
- Remove rigid 800px fixed-height BrowseServicesVirtualizedList
- Update next.config.mjs to redirect /services directly to /search?type=service
